requirements.txt
gensim==3.8.3
numpy>=1.23.0
textblob==0.15.3
nltk==3.5
streamlit==0.69.2
rake_nltk==1.0.4
spacy==2.3.2
spacy_streamlit==0.1.0
joblib==0.17.0
pandas==1.1.3
Pillow==8.3.2
scikit_learn==0.24.1
